server/pyproject.toml
[project]
name = "emma-server"
version = "0.1.0"
description = "Emma monitoring backend"
requires-python = ">=3.11"
authors = [{name = "Emma"}]
dependencies = [
  "fastapi~=0.110",
  "uvicorn[standard]~=0.29",
  "sqlalchemy[asyncio]~=2.0",
  "asyncpg~=0.29",
  "alembic~=1.13",
  "pydantic~=2.6",
  "pydantic-settings~=2.2",
  "pyjwt[crypto]~=2.8",
  "structlog~=23.2",
  "jinja2~=3.1",
  "python-multipart~=0.0.9",
  "httpx~=0.27"
]

[project.optional-dependencies]
dev = [
  "pytest~=8.1",
  "pytest-asyncio~=0.23",
  "pytest-cov~=4.1",
  "anyio~=4.2",
  "ruff~=0.4",
  "mypy~=1.8"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
